+++ tests/SearchableTest.php
@@ -4,7 +4,8 @@ namespace Laravel\Scout\Tests;
 
 use Mockery as m;
 use PHPUnit\Framework\TestCase;
-use Laravel\Scout\Tests\Fixtures\SearchableTestModel;
+use Illuminate\Database\Eloquent\Builder;
+use Laravel\Scout\Tests\Fixtures\SearchableModel;
 
 class SearchableTest extends TestCase
 {
@@ -19,7 +20,7 @@ class SearchableTest extends TestCase
         $collection->shouldReceive('isEmpty')->andReturn(false);
         $collection->shouldReceive('first->searchableUsing->update')->with($collection);
 
-        $model = new SearchableTestModel;
+        $model = new SearchableModel();
         $model->queueMakeSearchable($collection);
     }
 
@@ -29,7 +30,7 @@ class SearchableTest extends TestCase
         $collection->shouldReceive('isEmpty')->andReturn(true);
         $collection->shouldNotReceive('first->searchableUsing->update');
 
-        $model = new SearchableTestModel;
+        $model = new SearchableModel;
         $model->queueMakeSearchable($collection);
     }
 
@@ -39,7 +40,7 @@ class SearchableTest extends TestCase
         $collection->shouldReceive('isEmpty')->andReturn(false);
         $collection->shouldReceive('first->searchableUsing->delete')->with($collection);
 
-        $model = new SearchableTestModel;
+        $model = new SearchableModel;
         $model->queueRemoveFromSearch($collection);
     }
 
@@ -49,7 +50,7 @@ class SearchableTest extends TestCase
         $collection->shouldReceive('isEmpty')->andReturn(true);
         $collection->shouldNotReceive('first->searchableUsing->delete');
 
-        $model = new SearchableTestModel;
+        $model = new SearchableModel;
         $model->queueRemoveFromSearch($collection);
     }
 
@@ -59,11 +60,11 @@ class SearchableTest extends TestCase
     }
 }
 
-class ModelStubForMakeAllSearchable extends SearchableTestModel
+class ModelStubForMakeAllSearchable extends SearchableModel
 {
     public function newQuery()
     {
-        $mock = \Mockery::mock('Illuminate\Database\Eloquent\Builder');
+        $mock = m::mock(Builder::class);
 
         $mock->shouldReceive('orderBy')
             ->with('id')